+++ src/main/java/com/learnmore/legacy/domain/friends/presentation/dto/response/FriendRes.java
@@ -8,7 +8,7 @@ public record FriendRes(
         String nickname,
         String profileImage,
         Integer level,
-        Long styleId,
+        Integer styleId,
         String styleName,
         String friendCode,
         Boolean isKakaoFriend,
